Refactor MenuAppBar comp, emailjs tokens
