+++ .env
@@ -1,0 +1,2 @@
+EMAILJS_USER_ID="rohwedernt_emailjs"
+EMAILJS_TEMPLATE_ID="template_1fe4fhq"
