+++ src/views/MenuAppBar.jsx
@@ -1,44 +1,26 @@
-import React, { useState, Fragment } from 'react';
-import { Link } from "react-router-dom";
-import emailjs from 'emailjs';
-import apiKey from '../emailkey';
+import React, { useState } from 'react';
 
 // @material-ui/core components
 import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
-import IconButton from '@material-ui/core/IconButton';
-import MenuItem from '@material-ui/core/MenuItem';
-import Menu from '@material-ui/core/Menu';
-import { Divider } from '@material-ui/core';
-import Slide from '@material-ui/core/Slide';
-import useScrollTrigger from '@material-ui/core/useScrollTrigger';
-
-// @material-ui/icons
-import InfoIcon from '@material-ui/icons/Info';
-import MenuIcon from '@material-ui/icons/Menu';
-import EmailIcon from '@material-ui/icons/Email';
-import SettingsIcon from '@material-ui/icons/Settings';
-import HomeIcon from '@material-ui/icons/Home';
 
 // custom components
 import CustomDialogFullScreen from '../components/Dialogs/CustomDialogFullScreen';
 import CustomDialog from '../components/Dialogs/CustomDialog';
 import SuccessAlert from '../components/Alerts/SuccessAlert';
-import SettingsDialogContent from '../components/Dialogs/DialogContent/SettingsDialogContent';
-import EmailDialogContent from '../components/Dialogs/DialogContent/EmailDialogContent';
-import { AboutThisSite } from './Content/AboutThisSite';
-import { ComingSoon } from './Content/ComingSoon';
-import AstronomyContainer from './Containers/AstronomyContainer';
+import HamburgerMenu from './MenuSections/HamburgerMenu';
+import MenuIconButtons from './MenuSections/MenuIconButtons';
+import HideOnScroll from '../components/Animators/HideOnScroll';
 //import { Nrdcv1 } from 'nrdcv1';
 
 
 const useStyles = makeStyles((theme) => ({
   root: {
-	flexGrow: 1,
-	'& .MuiIconButton-root:hover': {
-		backgroundColor: 'rgba(0,0,0,0.2)'
-	}
+    flexGrow: 1,
+    '& .MuiIconButton-root:hover': {
+      backgroundColor: 'rgba(0,0,0,0.2)'
+    }
   },
   menuButton: {
     marginRight: theme.spacing(2),
@@ -52,116 +34,23 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-function HideOnScroll(props) {
-  const trigger = useScrollTrigger();
-
-  return (
-    <Slide appear={false} in={!trigger}>
-      {props.children}
-    </Slide>
-  );
-}
-
 export default function MenuAppBar(props) {
   const classes = useStyles();
-  const { breakpointMd, breakpointSm, handleThemeChange, theme, themeColor, setThemeColor, themeType, setThemeType, onAdmin } = props;
-
-  // success alert state
+  const { breakpointMd, breakpointSm } = props;
   const [openAlert, setOpenAlert] = useState(false);
-
-  // 'more stuff' popover menu
   const [anchorEl, setAnchorEl] = useState(null);
-  const openMenu = Boolean(anchorEl);
-  const handleMenu = (event) => setAnchorEl(event.currentTarget);
   const handleCloseMenu = () => setAnchorEl(null);
-
-  // regular dialog
   const [openDialog, setOpenDialog] = useState(false);
   const [dialogProps, setDialogProps] = useState({ title: '', description: '', content: () => {}});
+  const handleCloseDialog = () => setOpenDialog(false);
+  const [openDialogFull, setOpenDialogFull] = useState(false);
+  const [dialogFullProps, setDialogFullProps] = useState({ title: '', description: '', content: () => {}});
+  const handleCloseDialogFull = () => setOpenDialogFull(false);
+
   const handleOpenDialog = (data) => {
     setDialogProps(data);
     handleCloseMenu();
     setOpenDialog(true);
-  };
-  const handleCloseDialog = () => setOpenDialog(false);
-  const handleSubmitDialog = (e) => {
-      e.preventDefault();
-      emailjs.sendForm(`gmail`, apiKey.TEMPLATE_ID, e.target, apiKey.USER_ID)
-        .then((result) => {
-            alert("Message Sent, We will get back to you shortly", result.text);
-          },
-          (error) => {
-            alert("An error occurred, Please try again", error.text);
-        });
-    setOpenDialog(false);
-  }
-
-
-  // full screen dialog
-  const [openDialogFull, setOpenDialogFull] = useState(false);
-  const [dialogFullProps, setDialogFullProps] = useState({ title: '', description: '', content: () => {}});
-  const handleOpenDialogFull = (data) => {
-    setDialogFullProps(data);
-    handleCloseMenu();
-    setOpenDialogFull(true);
-  };
-  const handleCloseDialogFull = () => setOpenDialogFull(false);
-
-  // Menu item props
-  const aboutThisSiteProps = { 
-    content: () => <AboutThisSite />
-  };
-
-  const comingSoonProps = {
-    title: 'Features and Content Coming Soon',
-    content: () => <ComingSoon />
-  };
-
-  const nasaApiProps = {
-    title: 'NASA Globe API Example',
-    description: 'A simple rendering of the interactive globe available from the NASA API',
-    content: () => <AstronomyContainer />,
-    comingSoon: false
-  };
-
-  const scrumRetroToolProps = {
-    title: 'Scrum Retro Tool',
-    description: 'This is a from-scratch tool used for managing custom retrospective formats for scrum teams.',
-    content: () => {},
-    comingSoon: true
-  };
-
-  const dashboardProps = {
-    title: 'Dashboard POC',
-    description: 'This is a simple test app for showcasing some dashboard components I was interested in. Eventually I hope to hook this up to real data collected by AWS Kinesis.',
-    content: () => {},
-    comingSoon: true
-  };
-
-  const nrV1Props = {
-    title: 'NateRohwederDotCom v1',
-    description: 'Over time as the site evolves I want to be able to showcase past versions of the site. This was the first iteration, enjoy.',
-    content: () => {},
-    comingSoon: true
-  };
-
-  const contactProps = {
-    title: 'Contact',
-    confirmText: 'Send',
-    confirmFunc: () => handleSubmitDialog(),
-    content: () => <EmailDialogContent />
-  };
-
-  const settingsProps = {
-    title: 'Settings',
-    content: () => <SettingsDialogContent 
-      handleThemeChange={handleThemeChange}
-      theme={theme}
-      themeColor={themeColor}
-      setThemeColor={setThemeColor}
-      themeType={themeType}
-      setThemeType={setThemeType}
-    />
   };
 
   return (
@@ -169,70 +58,17 @@ export default function MenuAppBar(props) {
       <HideOnScroll {...props}>
         <AppBar>
           <Toolbar style={{ justifyContent: 'space-between'}}>
-            {onAdmin ? (
-              <IconButton
-                //component={Link}
-                to={'/'}
-                color='inherit'
-                className={classes.back}
-              >
-                <HomeIcon />
-              </IconButton>
-            ) : (
-              <Fragment>
-                <IconButton
-                  aria-label='more stuff'
-                  aria-controls='menu-appbar'
-                  aria-haspopup='true'
-                  onClick={handleMenu}
-                  color='secondary'
-                  className={classes.menuIcon}
-                >
-                  {/* <Typography style={{ marginRight: '10px' }} variant='subtitle1'>More Stuff</Typography> */}
-                  <MenuIcon />
-                </IconButton>
-                <Menu
-                  id='menu-appbar'
-                  anchorEl={anchorEl}
-                  anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-                  keepMounted
-                  transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-                  open={openMenu}
-                  onClose={handleCloseMenu}
-                >
-                  <MenuItem onClick={() => handleOpenDialogFull(nasaApiProps)}>NASA Interactive Globe</MenuItem>
-                  <MenuItem onClick={() => handleOpenDialogFull(scrumRetroToolProps)}>Scrum Retro Tool</MenuItem>
-                  <MenuItem onClick={() => handleOpenDialogFull(dashboardProps)}>Dashboard POC</MenuItem>
-                  <MenuItem onClick={() => handleOpenDialogFull(nrV1Props)}>NateRohwederDotCom v1</MenuItem>
-                  <Divider />
-                  <MenuItem onClick={() => handleOpenDialog(comingSoonProps)}>Coming Soon</MenuItem>
-                  {/* <MenuItem component={Link} to={'/admin'}>Admin</MenuItem> */}
-                </Menu>
-              </Fragment>
-            )}
-            {!onAdmin && <div>
-              <IconButton
-                onClick={() => handleOpenDialog(aboutThisSiteProps)}
-                color='secondary'
-                className={classes.menuIcon}
-              >
-                <InfoIcon />
-              </IconButton>
-              <IconButton
-                onClick={() => handleOpenDialog(contactProps)}
-                color='secondary'
-                className={classes.menuIcon}
-              >
-                <EmailIcon />
-              </IconButton>
-              <IconButton
-                onClick={() => handleOpenDialog(settingsProps)}
-                color='secondary'
-                className={classes.menuIcon}
-              >
-                <SettingsIcon />
-              </IconButton>
-            </div>}
+            <HamburgerMenu 
+              anchorEl={anchorEl}
+              setAnchorEl={setAnchorEl}
+              setDialogFullProps={setDialogFullProps}
+              handleCloseMenu={handleCloseMenu}
+              setOpenDialogFull={setOpenDialogFull}
+              handleOpenDialog={handleOpenDialog}
+            />
+            <MenuIconButtons 
+              handleOpenDialog={handleOpenDialog}
+            />
           </Toolbar>
         </AppBar>
       </HideOnScroll>
